# pkg/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for schedule, assessments, attendance and grades.",
        "title": "GradeFlow API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/assessments": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assessments"
                ],
                "summary": "List assessments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.AssessmentList"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assessments"
                ],
                "summary": "Create assessment",
                "parameters": [
                    {
                        "description": "assessment",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateAssessment"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Assessment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/assessments/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assessments"
                ],
                "summary": "Get assessment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "assessment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Assessment"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assessments"
                ],
                "summary": "Update assessment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "assessment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "assessment",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateAssessment"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Assessment"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assessments"
                ],
                "summary": "Delete assessment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "assessment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                }
            }
        },
        "/assessments/{id}/grades": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assessments",
                    "grades"
                ],
                "summary": "List grades for assessment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "assessment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.AssessmentGradeList"
                        }
                    }
                }
            }
        },
        "/assessments/{id}/grades/bulk": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "assessments",
                    "grades"
                ],
                "summary": "Bulk upsert grades for assessment",
                "parameters": [
                    {
                        "type": "string",
                        "description": "assessment id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "grades",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/request.GradeItem"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.AssessmentGradeList"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "request.CreateAssessment": {
            "type": "object",
            "required": [
                "courseId",
                "dateAt",
                "scale",
                "type"
            ],
            "properties": {
                "courseId": {
                    "type": "string"
                },
                "dateAt": {
                    "type": "string"
                },
                "maxPts": {
                    "type": "number"
                },
                "room": {
                    "type": "string"
                },
                "scale": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "request.GradeItem": {
            "type": "object",
            "required": [
                "studentId"
            ],
            "properties": {
                "scale": {
                    "type": "string"
                },
                "studentId": {
                    "type": "string"
                },
                "valueNum": {
                    "type": "number"
                },
                "valuePass": {
                    "type": "boolean"
                }
            }
        },
...
        "request.UpdateAssessment": {
            "type": "object",
            "properties": {
                "courseId": {
                    "type": "string"
                },
                "dateAt": {
                    "type": "string"
                },
                "maxPts": {
                    "type": "number"
                },
                "room": {
                    "type": "string"
                },
                "scale": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "response.Assessment": {
            "type": "object",
            "properties": {
                "courseId": {
                    "type": "string"
                },
                "dateAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "maxPts": {
                    "type": "number"
                },
                "room": {
                    "type": "string"
                },
                "scale": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "response.AssessmentGrade": {
            "type": "object",
            "properties": {
                "assessmentId": {
                    "type": "string"
                },
                "gradedAt": {
                    "type": "string"
                },
                "gradedBy": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "scale": {
                    "type": "string"
                },
                "studentId": {
                    "type": "string"
                },
                "valueNum": {
                    "type": "number"
                },
                "valuePass": {
                    "type": "boolean"
                }
            }
        },
        "response.AssessmentGradeList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.AssessmentGrade"
                    }
                },
                "page": {
                    "$ref": "#/definitions/response.Page"
                }
            }
        },
        "response.AssessmentList": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.Assessment"
                    }
                },
                "page": {
                    "$ref": "#/definitions/response.Page"
                }
            }
        },
...
        "response.Error": {
            "type": "object",
            "properties": {
                "details": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
        "response.Page": {
            "type": "object",
            "properties": {
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter the token with the `Bearer ` prefix, e.g. `Bearer \u003ctoken\u003e`",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}